Updated to use ProteinStructures datatype.
The ModelProteinStructure datatype has been deprecated in favor of the
ProteinStructures datatype. This datatype is still under active development, so
there may need to be more changes in the future to keep this app operational.

Additionally CompoundSetUtils does not support uploads outside of the staging
area at this time, so ligand upload is disabled.

This commit also updates the DataFileUtil and ProteinStructureUtils clients.