Updated Dockerfile to include node and app build.
The react app should be built when the docker image is built, not when the user
runs the KBase app. When the KBase app creates its report, it will create a
JSON file with the report data which will be rendered by the react app. This
data file and the previously built react app will be included into the KBase
app report and uploaded to the workspace.